refactor(site/pages/AgentsPage): extract ScrollAnchoredContainer into its own module
Extract the ~750-line ScrollAnchoredContainer component and its scroll
anchoring logic from AgentChatPageView.tsx (1291 lines) into a dedicated
module under components/ScrollAnchoredContainer/.

New files:
- useScrollAnchoring.ts: custom hook owning all refs, effects, and
  imperative scroll logic. The standalone scrollTranscriptToBottom
  function that previously took a bag of refs is now a closure inside
  the hook with a simplified (behavior) => void signature.
- ScrollAnchoredContainer.tsx: thin presentational shell (~80 lines)
  using the hook.
- index.ts: barrel export.

AgentChatPageView.tsx drops from 1291 to 531 lines, under the 500-line
guideline target. Zero behavior change; all scroll anchoring, sentinel
observation, resize handling, and touch/wheel guards are preserved
exactly.
