+++ site/src/pages/AgentsPage/components/ScrollAnchoredContainer/index.ts
@@ -1,0 +1,1 @@
+export { ScrollAnchoredContainer } from "./ScrollAnchoredContainer";
